+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/common/setup/RecipeSerializers.kt
@@ -8,7 +8,7 @@ import site.siredvin.peripheralworks.xplat.PeripheralWorksPlatform
 object RecipeSerializers {
     val STATUE_CLONING = PeripheralWorksPlatform.registerRecipeSerializer(
         modId("statue_cloning"),
-        SimpleCraftingRecipeSerializer(::StatueCloningRecipe)
+        SimpleCraftingRecipeSerializer(::StatueCloningRecipe),
     )
 
     fun doSomething() {}