+++ projects/fabric/src/main/kotlin/site/siredvin/peripheralworks/client/FlexibleRealityAnchorModel.kt
@@ -9,7 +9,6 @@ import net.minecraft.client.Minecraft
 import net.minecraft.client.renderer.block.model.BakedQuad
 import net.minecraft.client.renderer.block.model.ItemOverrides
 import net.minecraft.client.renderer.block.model.ItemTransforms
-import net.minecraft.client.renderer.texture.TextureAtlas
 import net.minecraft.client.renderer.texture.TextureAtlasSprite
 import net.minecraft.client.resources.model.*
 import net.minecraft.core.BlockPos