+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Statue workbench
+- Alloy forgery integration
+
 ## [1.0.0] - 2023-07-09
 
 ### Added